Refactor tweet tile content & replies
